кнопки
попытки сделать так, что бы кнопки пропадали и снова появлялись

+++ Assets/Scripts/OnButtonHover.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class OnButtonHover : MonoBehaviour
 {
-    public int AmountToAdd;
+    public ulong AmountToAdd;
 
     public void ifHoveredOn()
     {
